+++ src/app/_component/banner/Banner.tsx
