+++ src/app/_component/banner/CarouselBanner.tsx
@@ -2,7 +2,7 @@
 
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { Carousel } from 'react-responsive-carousel';
-import { CAROUSEL } from '@lib/const/carousel.ts';
+import { CAROUSEL } from '@lib/const/carousel';
 import throttle from '@lib/utils/throttle';
 import * as styles from './banner.css';
 import 'react-responsive-carousel/lib/styles/carousel.min.css';
