+++ src/app/_component/banner/banner.css.ts
